Remove possible duplicate code
