+++ .idea/workspace.xml
@@ -2,8 +2,11 @@
 <project version="4">
   <component name="ChangeListManager">
     <list default="true" id="47758b3c-adb1-47d1-96bb-f0611aac81b4" name="Default Changelist" comment="">
-      <change beforePath="$PROJECT_DIR$/src/pages/sales-invoice/new/new.html" beforeDir="false" afterPath="$PROJECT_DIR$/src/pages/sales-invoice/new/new.html" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/pages/sales-invoice/new/new.ts" beforeDir="false" afterPath="$PROJECT_DIR$/src/pages/sales-invoice/new/new.ts" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/pages/customer/customers/customers.ts" beforeDir="false" afterPath="$PROJECT_DIR$/src/pages/customer/customers/customers.ts" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/pages/purchase-invoice/new/new.ts" beforeDir="false" afterPath="$PROJECT_DIR$/src/pages/purchase-invoice/new/new.ts" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/pages/purchase-invoice/overview/overview.ts" beforeDir="false" afterPath="$PROJECT_DIR$/src/pages/purchase-invoice/overview/overview.ts" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/pages/sales-invoice/overview/overview.ts" beforeDir="false" afterPath="$PROJECT_DIR$/src/pages/sales-invoice/overview/overview.ts" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/util/time-converter.ts" beforeDir="false" afterPath="$PROJECT_DIR$/src/util/time-utils.ts" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -96,7 +99,7 @@
       <workItem from="1586105274645" duration="8072000" />
       <workItem from="1586671491488" duration="6697000" />
       <workItem from="1586706226511" duration="398000" />
-      <workItem from="1586772675385" duration="1535000" />
+      <workItem from="1586772675385" duration="2033000" />
     </task>
     <task id="LOCAL-00001" summary="Move frontend to docker for easier nginx management">
       <created>1584006004551</created>
@@ -334,7 +337,14 @@
       <option name="project" value="LOCAL" />
       <updated>1586773999175</updated>
     </task>
-    <option name="localTasksCounter" value="35" />
+    <task id="LOCAL-00035" summary="Add product select for sales invoice">
+      <created>1586774505702</created>
+      <option name="number" value="00035" />
+      <option name="presentableId" value="LOCAL-00035" />
+      <option name="project" value="LOCAL" />
+      <updated>1586774505702</updated>
+    </task>
+    <option name="localTasksCounter" value="36" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
@@ -353,7 +363,6 @@
     <option name="oldMeFiltersMigrated" value="true" />
   </component>
   <component name="VcsManagerConfiguration">
-    <MESSAGE value="Add running instructions" />
     <MESSAGE value="Move aurelia to port 6060 to avoid conflicts with ofbiz during development" />
     <MESSAGE value="Added header, refactored code, added bootstrap" />
     <MESSAGE value="Add json data" />
@@ -378,7 +387,8 @@
     <MESSAGE value="Add billing account service" />
     <MESSAGE value="Add account selection" />
     <MESSAGE value="More consistent naming" />
-    <option name="LAST_COMMIT_MESSAGE" value="More consistent naming" />
+    <MESSAGE value="Add product select for sales invoice" />
+    <option name="LAST_COMMIT_MESSAGE" value="Add product select for sales invoice" />
   </component>
   <component name="WindowStateProjectService">
     <state x="677" y="357" width="654" height="423" key="#com.intellij.fileTypes.FileTypeChooser" timestamp="1586673211684">
@@ -390,11 +400,11 @@
       <screen x="70" y="27" width="1850" height="1053" />
     </state>
     <state x="803" y="272" width="374" height="569" key="#com.intellij.ide.util.MemberChooser/70.27.1850.1053@70.27.1850.1053" timestamp="1586773726706" />
-    <state x="539" y="91" width="1000" height="989" key="CommitChangelistDialog2" timestamp="1586773999115">
+    <state x="539" y="91" width="1000" height="989" key="CommitChangelistDialog2" timestamp="1586774505650">
       <screen x="70" y="27" width="1850" height="1053" />
     </state>
     <state x="537" y="149" width="903" height="807" key="CommitChangelistDialog2/67.27.1853.1053@67.27.1853.1053" timestamp="1584120291815" />
-    <state x="539" y="91" width="1000" height="989" key="CommitChangelistDialog2/70.27.1850.1053@70.27.1850.1053" timestamp="1586773999115" />
+    <state x="539" y="91" width="1000" height="989" key="CommitChangelistDialog2/70.27.1850.1053@70.27.1850.1053" timestamp="1586774505650" />
     <state width="1826" height="265" key="GridCell.Tab.0.bottom" timestamp="1586706057949">
       <screen x="70" y="27" width="1850" height="1053" />
     </state>
@@ -419,11 +429,11 @@
       <screen x="70" y="27" width="1850" height="1053" />
     </state>
     <state x="484" y="188" width="1022" height="737" key="SettingsEditor/70.27.1850.1053@70.27.1850.1053" timestamp="1586671603499" />
-    <state x="590" y="288" width="800" height="543" key="Vcs.Push.Dialog.v2" timestamp="1586774002392">
+    <state x="590" y="288" width="800" height="543" key="Vcs.Push.Dialog.v2" timestamp="1586774507232">
       <screen x="70" y="27" width="1850" height="1053" />
     </state>
     <state x="588" y="288" width="800" height="528" key="Vcs.Push.Dialog.v2/67.27.1853.1053@67.27.1853.1053" timestamp="1584120294550" />
-    <state x="590" y="288" width="800" height="543" key="Vcs.Push.Dialog.v2/70.27.1850.1053@70.27.1850.1053" timestamp="1586774002392" />
+    <state x="590" y="288" width="800" height="543" key="Vcs.Push.Dialog.v2/70.27.1850.1053@70.27.1850.1053" timestamp="1586774507232" />
     <state x="649" y="380" key="com.intellij.ide.util.TipDialog" timestamp="1584094321552">
       <screen x="67" y="27" width="1853" height="1053" />
     </state>
