+++ src/pages/customer/customers/customers.ts
@@ -1,6 +1,6 @@
 import {PartyService} from "../../../service/party-service";
 import {inject} from "aurelia-dependency-injection";
-import {TimeConverter} from "../../../util/time-converter";
+import {TimeUtils} from "../../../util/time-utils";
 
 @inject(PartyService)
 export class Customers {
@@ -16,7 +16,7 @@ export class Customers {
     }
 
     convertTime(ms: number) {
-        return TimeConverter.convertDate(ms);
+        return TimeUtils.convertDate(ms);
     }
 
 }
